fuelstations/requirements.txt
asgiref==3.3.4
attrs==21.2.0
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
dateparser==1.0.0
fastapi==0.65.1
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-bigquery==2.19.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httptools==0.2.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.5.0
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
packaging==20.9
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
proto-plus==1.18.1
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyspark==3.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
starlette==0.14.2
Twisted==21.2.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
w3lib==1.22.0
watchgod==0.7
websockets==9.1
zipp==3.4.1
zope.interface==5.4.0
